+++ src/main/java/net/t7seven7t/craftfx/effect/Potion.java
@@ -1,10 +1,11 @@
 /**
  * Copyright (C) 2013 t7seven7t
  */
-package net.t7seven7t.craftfx.effects;
+package net.t7seven7t.craftfx.effect;
 
-import net.t7seven7t.craftfx.ItemLoader;
 import net.t7seven7t.craftfx.Trigger;
+import net.t7seven7t.craftfx.item.ItemLoader;
+
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.inventory.ItemStack;
@@ -20,13 +21,13 @@ public class Potion extends Effect {
 	List<PotionEffect> potionEffects;
 	
 	public Potion(Trigger trigger, ItemStack item, ConfigurationSection config) throws Exception {
-		super(EffectType.POTION, trigger, item);
+		super(trigger, item);
 		
 		this.potionEffects = ItemLoader.getPotionEffects(config.getStringList("potion-effects"));
 	}
 	
 	public Potion(Trigger trigger, ItemStack item, List<PotionEffect> potionEffects) {
-		super(EffectType.POTION, trigger, item);
+		super(trigger, item);
 		
 		this.cancelsAction = true;
 		this.potionEffects = potionEffects;