+++ src/main/java/net/t7seven7t/craftfx/recipe/FXShapedRecipe.java
@@ -1,7 +1,9 @@
 /**
  * Copyright (C) 2013 t7seven7t
  */
-package net.t7seven7t.craftfx;
+package net.t7seven7t.craftfx.recipe;
+
+import net.t7seven7t.craftfx.CraftFX;
 
 import org.apache.commons.lang.Validate;
 import org.bukkit.inventory.ItemStack;