+++ src/main/java/net/t7seven7t/craftfx/recipe/FXShapelessRecipe.java
@@ -1,9 +1,12 @@
 /**
  * Copyright (C) 2013 t7seven7t
  */
-package net.t7seven7t.craftfx;
+package net.t7seven7t.craftfx.recipe;
 
 import com.google.common.collect.Lists;
+
+import net.t7seven7t.craftfx.CraftFX;
+
 import org.apache.commons.lang.Validate;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.ShapelessRecipe;