+++ CHANGES.md
@@ -1,6 +1,12 @@
 CHANGES
 =======
 
+Version 0.0.2
+-------------
+- Moved classes around for better package structure (new stuff is permanent now)
+- Added custom effect class loader so you can write your own effects easily :)
+- Updated silly old config items
+
 Version 0.0.1
 -------------
 - Initial release